+++ src/classes/babyClasses.ts
@@ -223,6 +223,7 @@ export { OrangeGhostBaby } from "./babies/OrangeGhostBaby";
 export { OrangePigBaby } from "./babies/OrangePigBaby";
 export { PandaBaby } from "./babies/PandaBaby";
 export { PegasusBaby } from "./babies/PegasusBaby";
+export { PenguinBaby } from "./babies/PenguinBaby";
 export { PieceABaby } from "./babies/PieceABaby";
 export { PieceBBaby } from "./babies/PieceBBaby";
 export { PinkGhostBaby } from "./babies/PinkGhostBaby";